requirements.txt
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
dnspython==2.2.1
Flask==2.2.2
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
keras==2.10.0
# Keras-Preprocessing==1.1.2
MarkupSafe==2.1.1
numpy==1.23.1
pandas==1.5.0
pycparser==2.21
PyJWT==2.4.0
pymongo==4.2.0
python-dotenv==0.20.0
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
sklearn==0.0
tensorflow==2.10.0
threadpoolctl==3.1.0
Werkzeug==2.2.2